requirements.txt
nltk
ipykernal
scipy 
numpy
cupy-cuda12x
scikit-learn
imbalanced-learn
tensorflow
torch 
torchvision 
torchaudio


